+++ src/test/java/com/arangodb/ArangoCollectionTest.java
@@ -972,7 +972,7 @@ public class ArangoCollectionTest extends BaseTest {
 		assertThat(indexResult.getIsNewlyCreated(), is(true));
 		assertThat(indexResult.getMinLength(), is(nullValue()));
 		if (arangoDB.getRole() == ServerRole.SINGLE) {
-			assertThat(indexResult.getSelectivityEstimate(), is(1));
+			assertThat(indexResult.getSelectivityEstimate(), is(1.));
 		}
 		assertThat(indexResult.getSparse(), is(false));
 		assertThat(indexResult.getType(), is(IndexType.hash));