+++ ChangeLog.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 
 - changed the internal connection pool and host management. There now exists a connection pool for every configured host. This changes the behavior of `ArangoDB.Builder#maxConnections` which now allows to configure the maximal number of connection per host and not overall.
+- changed `IndexEntity#selectivityEstimate` from `Integer` to `Double`
 
 ## [4.7.3] - 2018-09-03